requirements.txt
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
pytz>=2023.3
python-dotenv>=1.0.0
playwright>=1.40.0  # 用于微博自动化发布
